refactor: 커맨드 클래스 패키지 수정
