+++ src/main/java/org/prgrms/kdt/KdtApplication.java
@@ -1,7 +1,7 @@
 package org.prgrms.kdt;
 
-import org.prgrms.kdt.domain.command.types.CommandType;
-import org.prgrms.kdt.domain.command.componet.CommandComponent;
+import org.prgrms.kdt.domain.command.CommandType;
+import org.prgrms.kdt.domain.command.CommandProcessor;
 import org.prgrms.kdt.util.Input;
 import org.prgrms.kdt.util.Output;
 import org.springframework.boot.SpringApplication;
@@ -15,13 +15,13 @@ public class KdtApplication {
 
 	public static void main(String[] args) {
 		ConfigurableApplicationContext context = SpringApplication.run(KdtApplication.class, args);
-		CommandComponent commandComponent = context.getBean(CommandComponent.class);
+		CommandProcessor commandProcessor = context.getBean(CommandProcessor.class);
 		CommandType commandType;
 
 		do {
 			Output.printVoucherProgram();
 			commandType = Input.inputVoucherCommand();
-			commandComponent.processCommand(commandType);
+			commandProcessor.doCommand(commandType);
 		} while (commandType != CommandType.EXIT);
 	}
 }
