+++ src/test/java/org/prgrms/kdt/domain/command/types/CommandTypeTest.java
@@ -2,8 +2,8 @@ package org.prgrms.kdt.domain.command.types;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EnumSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import org.prgrms.kdt.domain.command.CommandType;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
